requirements.txt
Django==1.11.1
psycopg2==2.7.1

Markdown==2.6.8
django-filter==1.0.4
djangorestframework==3.6.3
djangorestframework-gis==0.11.2

# Development tools
ipdb
